pyproject.toml
[project]
name = "at-ai-editor-recommender"
version = "1.2.10"
description = "AI editor recommender"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "aioboto3>=7.0.0",
    "aiohttp>=3.12.15",
    "boto3>=1.40.6",
    "fastapi>=0.116.1",
    "psycopg[binary]>=3.2.0",
    "uvicorn>=0.35.0",
    "jinja2>=3.1.6",
    "ddtrace>=2.16.0",  # Datadog tracing
    # AWS Strands Agents SDK
    "strands-agents>=1.29.0",
    # Long-term memory (Tier 3) — Postgres store
    "langgraph-checkpoint-postgres>=2.0.0",
]

[project.scripts]
ee-workflow = "at_ai_editor_recommender.ee_workflow_langgraph_otel:main"

[build-system]
requires = ["uv_build>=0.8,<0.9"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.3.7",
    "streamlit>=1.48.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "psycopg[binary]>=3.2.0",
]
